373 - client - Update stores
