+++ client/src/pages/automation/project/stores/useConnectionNoteStore.ts
@@ -1,0 +1,24 @@
+import {create} from 'zustand';
+import {devtools, persist} from 'zustand/middleware';
+
+interface ConnectionNoteState {
+    showConnectionNote: boolean;
+    setShowConnectionNote: (showConnectionNote: boolean) => void;
+}
+
+export const useConnectionNoteStore = create<ConnectionNoteState>()(
+    devtools(
+        persist(
+            (set) => ({
+                setShowConnectionNote: (connectionNoteStatus) =>
+                    set(() => ({
+                        showConnectionNote: connectionNoteStatus,
+                    })),
+                showConnectionNote: true,
+            }),
+            {
+                name: 'connection-note',
+            }
+        )
+    )
+);
